+++ src/pages/HomePage.js
@@ -1,16 +1,14 @@
 import React from 'react'
 import WelcomeUser from '../components/HomePage/WelcomeUser'
-import NavButton from '../components/HomePage/NavButton'
+import NavButtons from '../components/HomePage/NavButtons'
 import TopLeftCard from '../components/Design/CardWrapper/TopLeftCard'
-import HeaderHomepage from '../components/Header/HeaderHomepage'
 
 export default function HomePage({ user }) {
   return (
     <>
-      <HeaderHomepage />
       <TopLeftCard />
       <WelcomeUser user={user} />
-      <NavButton />
+      <NavButtons />
     </>
   )
 }
